requirements.txt
Flask==2.0.3
pandas==1.3.3
python-docx==0.8.11
openpyxl==3.0.9
pypandoc==1.14